Move Extensions and TestArchTaskExecutor for test into testutil module.
