+++ testutil/src/main/java/jp/kuaddo/tsuidezake/testutil/Extensions.kt
@@ -1,5 +1,6 @@
-package jp.kuaddo.tsuidezake
+package jp.kuaddo.tsuidezake.testutil
 
+import androidx.annotation.VisibleForTesting
 import androidx.arch.core.executor.ArchTaskExecutor
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.Observer
@@ -11,6 +12,7 @@ import kotlinx.coroutines.test.resetMain
 import kotlinx.coroutines.test.setMain
 import org.spekframework.spek2.dsl.GroupBody
 
+@VisibleForTesting(otherwise = VisibleForTesting.NONE)
 fun GroupBody.applyArchTaskExecutor() {
 
     beforeEachTest { ArchTaskExecutor.getInstance().setDelegate(TestArchTaskExecutor()) }
@@ -18,6 +20,7 @@ fun GroupBody.applyArchTaskExecutor() {
     afterEachTest { ArchTaskExecutor.getInstance().setDelegate(null) }
 }
 
+@VisibleForTesting(otherwise = VisibleForTesting.NONE)
 @ExperimentalCoroutinesApi
 fun GroupBody.applyTestDispatcher(dispatcher: CoroutineDispatcher = Dispatchers.Unconfined) {
     beforeEachTest { Dispatchers.setMain(dispatcher) }
@@ -25,4 +28,5 @@ fun GroupBody.applyTestDispatcher(dispatcher: CoroutineDispatcher = Dispatchers.
     afterEachTest { Dispatchers.resetMain() }
 }
 
+@VisibleForTesting(otherwise = VisibleForTesting.NONE)
 fun <T> LiveData<T>.observeAndGet() = mockk<Observer<T>>(relaxed = true).also { observeForever(it) }
