+++ testutil/src/main/java/jp/kuaddo/tsuidezake/testutil/TestArchTaskExecutor.kt
@@ -1,7 +1,9 @@
-package jp.kuaddo.tsuidezake
+package jp.kuaddo.tsuidezake.testutil
 
+import androidx.annotation.VisibleForTesting
 import androidx.arch.core.executor.TaskExecutor
 
+@VisibleForTesting(otherwise = VisibleForTesting.NONE)
 class TestArchTaskExecutor : TaskExecutor() {
     override fun executeOnDiskIO(runnable: Runnable) {
         runnable.run()
